Add validation logic to PlaceController and UserController

+++ src/main/kotlin/hu/zsof/restaurantApp/controller/PlaceController.kt
@@ -4,29 +4,25 @@ import hu.zsof.restaurantApp.dto.PlaceDto
 import hu.zsof.restaurantApp.model.Place
 import hu.zsof.restaurantApp.model.convertToDto
 import hu.zsof.restaurantApp.service.PlaceService
+import hu.zsof.restaurantApp.util.AuthUtils
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 import java.util.*
 
 @RestController
 @RequestMapping("/places")
 class PlaceController(private val placeService: PlaceService) {
 
-    @PostMapping()
-    fun newPlace(@RequestBody place: Place): ResponseEntity<Place> {
-        val newPlace = placeService.newPlace(place)
-        return ResponseEntity(newPlace, HttpStatus.CREATED)
-    }
-
     @GetMapping("/{id}")
-    fun getPlaceById(@PathVariable id: Long): ResponseEntity<PlaceDto?> {
+    fun getPlaceById(
+        @PathVariable id: Long,
+        @CookieValue(AuthUtils.COOKIE_NAME) token: String?
+    ): ResponseEntity<PlaceDto?> {
+        val verification = AuthUtils.verifyToken(token)
+        if (!verification.verified || !verification.isAdmin) {
+            return ResponseEntity(HttpStatus.UNAUTHORIZED)
+        }
         val place: Optional<Place> = placeService.getPlaceById(id)
         if (!place.isPresent) {
             return ResponseEntity(null, HttpStatus.BAD_REQUEST)
@@ -35,20 +31,35 @@ class PlaceController(private val placeService: PlaceService) {
     }
 
     @GetMapping
-    fun getAllPlace(): ResponseEntity<List<PlaceDto>> {
+    fun getAllPlace(@CookieValue(AuthUtils.COOKIE_NAME) token: String?): ResponseEntity<List<PlaceDto>> {
+        val verification = AuthUtils.verifyToken(token)
+        if (!verification.verified || !verification.isAdmin) {
+            return ResponseEntity(HttpStatus.UNAUTHORIZED)
+        }
         val places: MutableList<Place> = placeService.getAllPlace()
         return ResponseEntity<List<PlaceDto>>(places.convertToDto(), HttpStatus.OK)
     }
 
     @DeleteMapping("/{id}")
-    fun deleteById(@PathVariable id: Long): ResponseEntity<HttpStatus> {
+    fun deleteById(
+        @PathVariable id: Long,
+        @CookieValue(AuthUtils.COOKIE_NAME) token: String?
+    ): ResponseEntity<HttpStatus> {
+        val verification = AuthUtils.verifyToken(token)
+        if (!verification.verified || !verification.isAdmin) {
+            return ResponseEntity(HttpStatus.UNAUTHORIZED)
+        }
         return if (placeService.deletePlaceById(id)) {
             ResponseEntity(HttpStatus.OK)
         } else ResponseEntity(HttpStatus.NOT_FOUND)
     }
 
     @DeleteMapping
-    fun deleteAllPlace(): ResponseEntity<HttpStatus> {
+    fun deleteAllPlace(@CookieValue(AuthUtils.COOKIE_NAME) token: String?): ResponseEntity<HttpStatus> {
+        val verification = AuthUtils.verifyToken(token)
+        if (!verification.verified || !verification.isAdmin) {
+            return ResponseEntity(HttpStatus.UNAUTHORIZED)
+        }
         placeService.deleteAll()
         return ResponseEntity(HttpStatus.OK)
     }
